+++ src/main/kotlin/dev/protsenko/securityLinter/kubernetes/quickfix/ReplaceValueTo1000QuickFix.kt
@@ -1,0 +1,24 @@
+package dev.protsenko.securityLinter.kubernetes.quickfix
+
+import com.intellij.codeInsight.intention.preview.IntentionPreviewInfo
+import com.intellij.codeInspection.LocalQuickFix
+import com.intellij.codeInspection.ProblemDescriptor
+import com.intellij.codeInspection.util.IntentionFamilyName
+import com.intellij.openapi.project.Project
+import dev.protsenko.securityLinter.utils.PsiElementGenerator
+import org.jetbrains.yaml.psi.YAMLScalar
+
+class ReplaceValueTo1000QuickFix(@IntentionFamilyName private val message: String) : LocalQuickFix {
+    override fun getFamilyName(): @IntentionFamilyName String = this.message
+
+    override fun generatePreview(project: Project, previewDescriptor: ProblemDescriptor): IntentionPreviewInfo =
+        IntentionPreviewInfo.EMPTY
+
+    override fun applyFix(project: Project, descriptor: ProblemDescriptor) {
+        val psiElement = descriptor.psiElement as? YAMLScalar ?: return
+
+        psiElement.replace(
+            PsiElementGenerator.rawText(project, "1000") ?: return
+        )
+    }
+}