+++ src/main/kotlin/dev/protsenko/securityLinter/kubernetes/utils/PsiUtils.kt
@@ -1,0 +1,9 @@
+package dev.protsenko.securityLinter.kubernetes.utils
+
+import org.jetbrains.yaml.psi.YAMLQuotedText
+
+fun YAMLQuotedText.unquoted(): String = if (this.isSingleQuote){
+    this.text.trim('\'')
+} else {
+    this.text.trim('"')
+}